refactor: reorganize prefs context and update imports
